% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{clean_headers}
\alias{clean_headers}
\title{helper function for eq_clean_data}
\usage{
clean_headers(data)
}
\description{
helper function for eq_clean_data
}
\seealso{
\code{\link{eq_clean_data}}
}
\keyword{internal}
