% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomlabel.R
\name{compute_timeline_label}
\alias{compute_timeline_label}
\title{Function to compute data to be consumed by geom/stat_timelabel}
\usage{
compute_timeline_label(data, scales, n_max = NA, ...)
}
\arguments{
\item{data}{data frame passed from stat / geom}

\item{scales}{passed from stat/ geom}

\item{xmin}{date passed from stat / geom,  specifying the minimum date to be
considered}

\item{xmax}{date pased from stat /geom, specifying the maximum date to be
considered}
}
\value{
data frame to be consumed by geom
}
\description{
The function filters the data according to date. It returns data that are in
between xmin and xmax. Note that xmin and xmax can either be mapped (using
aes(xmin=???)) or being an argument to geom_timeline. If both are specified,
the mapped value will be considered. In addition, if n_max and byCol are
specified, only n_max rows orderd by byCol are returned. If no byCol is
specified, it returns n_max rows randomly.
}
