% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{draw_group_timeline}
\alias{draw_group_timeline}
\title{helper function for ggproto_timeline}
\usage{
draw_group_timeline(data, panel_params, coord)
}
\description{
helper function for ggproto_timeline
}
\seealso{
\code{\link{ggproto_timeline}}
}
\keyword{internal}
