% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{draw_group_timeline_label}
\alias{draw_group_timeline_label}
\title{helper function for ggproto_timeline_label}
\usage{
draw_group_timeline_label(data, panel_params, coord)
}
\description{
helper function for ggproto_timeline_label
}
\seealso{
\code{\link{ggproto_timeline_label}}
}
