% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtimeline.R
\name{draw_timeline}
\alias{draw_timeline}
\title{Function to prepare timeline grid objects}
\usage{
draw_timeline(data, panel_scales, coord, na.rm = FALSE)
}
\arguments{
\item{data}{being processed data for plotting}

\item{panel_scales}{a list containing the details of x-y scales}

\item{coord}{coord object}

\item{na.rm}{default param inherited from class}
}
\value{
grid objects for drawing
}
\description{
Note that this function is the draw_panel function of  GeomTime. It is taken
out as a separate function for easy visualisation purpose. The function
prepares the axis and the points of the plot.
}
