% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomlabel.R
\name{draw_timeline_label}
\alias{draw_timeline_label}
\title{Function to prepare timeline annotation objects}
\usage{
draw_timeline_label(data, panel_scales, coord, parse = FALSE, na.rm = FALSE,
  check_overlap = FALSE)
}
\arguments{
\item{data}{being processed data for plotting}

\item{panel_scales}{a list containing the details of x-y scales}

\item{coord}{coord object}

\item{parse}{boolean to indicate whether the text should be parsed or not}

\item{na.rm}{default param inherited from class}

\item{check_overlap}{default param inherited from class}
}
\value{
grid objects for drawing
}
\description{
Note that this function is the draw_panel function of GeomTimelabel. It is
taken out as a separate function for easy visualisation purpose. The function
prepares the vertical lines and text objects of the points.
}
