% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelineLabel.R
\name{draw_timelinelabel}
\alias{draw_timelinelabel}
\title{An helper function for \code{draw_panel} of geomTimelineLabel.}
\usage{
draw_timelinelabel(data, panel_scales, coord)
}
\arguments{
\item{data}{data}

\item{panel_scales}{panel_scales}

\item{coord}{coord}
}
\description{
An helper function for \code{draw_panel} of geomTimelineLabel.
}
