% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{empty_if_na}
\alias{empty_if_na}
\title{helper function for eq_create_label}
\usage{
empty_if_na(x, FUN, ...)
}
\description{
helper function for eq_create_label
}
\seealso{
\code{\link{eq_create_label}}
}
