% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create label text for eq_map()}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{(\emph{\link[dplyr]{tbl_df}}) noaa dataset cleaned with \code{\link{eq_clean_data}}}
}
\description{
Create a label text that can be used with \code{\link{eq_map}} by combining \emph{LOCATION}, \emph{MAGNITUDE} and \emph{TOTAL_DEATHS} into a html-string.
}
\examples{
\dontrun{
system.file("extdata", "noaa_earthquakes.tsv", package = "noaa") \%>\%
  eq_read_data \%>\%
  eq_clean_data \%>\%
  mutate(label_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "label_text")
}

}
\seealso{
\code{\link{eq_map}}
}
