% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Visualize earthquake locations on map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{(\emph{\link[dplyr]{tbl_df}}) noaa dataset cleaned with \code{\link{eq_clean_data}}}

\item{annot_col}{(\emph{character}) column to be used as labels}
}
\description{
Draw earthquake locations from noaa dataset on leaflet map.
}
\examples{
\dontrun{
system.file("extdata", "noaa_earthquakes.tsv", package = "noaa") \%>\%
  eq_read_data \%>\%
  eq_clean_data \%>\%
  eq_map(annot_col = "DATE")
}

}
\seealso{
\code{\link{eq_create_label}}
}
