% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_data.R
\name{eq_noaa_example}
\alias{eq_noaa_example}
\title{load example noaa data}
\usage{
eq_noaa_example()
}
\value{
example noaa data, that comes with noaa package, is returned as tibble (see \code{\link[dplyr]{tbl_df}})
}
\description{
Load raw noaa example dataset from U.S. National Oceanographic and Atmospheric Administration (NOAA): \cr
National Geophysical Data Center / World Data Service (NGDC/WDS): NCEI/WDS Global Significant \cr
Earthquake Database. NOAA National Centers for Environmental Information. doi:10.7289/V5TD9V7K
}
\examples{
\dontrun{
noaa_raw <- eq_noaa_example()
}

}
\references{
Earthquake Database. NOAA National Centers for Environmental Information. \cr
(\href{https://www.ncei.noaa.gov/access/metadata/landing-page/bin/iso?id=gov.noaa.ngdc.mgg.hazards:G012153}{doi:10.7289/V5TD9V7K})
}
\seealso{
\code{\link{eq_read_data}}
}
