% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{<title>}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = 5,
  ...
)
}
\arguments{
\item{<x>}{(\emph{character}) <description>.}
}
\value{
FIBS data is returned as tibble (see \code{\link[dplyr]{tbl_df}})
}
\description{
<description> \cr
\code{<function name>}.
}
\details{
<details> \code{\link{<other function name>}}
}
\examples{
\dontrun{
...
}

}
\references{
US National Highway Traffic Safety Administration \cr
(\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\seealso{
\code{\link{<other function name>}}
}
