% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomlabel.R
\docType{data}
\name{GeomTimelabel}
\alias{GeomTimelabel}
\title{Geom class for the annotation of Earthquake data}
\format{An object of class \code{GeomTimelabel} (inherits from \code{GeomLabel}, \code{Geom}, \code{ggproto}) of length 4.}
\usage{
GeomTimelabel
}
\value{
ggproto object
}
\description{
Geom class to prepare the text objects for the earthquake data. It inherits
from GeomLabel, and aims to prepare the vertical lines and annotation of the
earthquake data. For input arguments, please refer to
\code{ggplot2::GeomLabel} as the parent for further details.
}
\keyword{datasets}
