% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtimeline.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Geom class for the Earthquake data}
\format{An object of class \code{GeomTimeline} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}) of length 4.}
\usage{
GeomTimeline
}
\value{
ggproto object
}
\description{
Geom class to prepare the drawing objects for the earthquake data. It
inherits from GeomPoint, and aims to prepare the axis and points of the
earthquake data. For input arguments, please refer to
\code{ggplot2::GeomPoint} as the parent for further details.
}
\keyword{datasets}
