% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tools.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{Title}
\format{An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimelineLabel
}
\arguments{
\item{coord}{}
}
\description{
Title
}
\keyword{datasets}
