% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{get_timeline}
\alias{get_timeline}
\title{Get Timeline}
\usage{
get_timeline(data, countries = "*", xmin, xmax)
}
\arguments{
\item{data}{A data.frame with NOAA earthquake data}

\item{countries}{A vector of countries}

\item{xmin}{A String to represent minimum time value in timeline}

\item{xmax}{A String to represent maximum time value in timeline}
}
\description{
A wrapper function to generate visualization based on \code{countries} parameter.
}
\examples{
\dontrun{
get_timeline(noaa.df, "*","2000-01-01","2010-01-01")
get_timeline(noaa.df, c("CHINA", "INDIA"),"2000-01-01","2010-01-01")

}

}
