% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\docType{data}
\name{ggproto_timeline}
\alias{ggproto_timeline}
\title{ggproto function for geom_timeline}
\format{
An object of class \code{ggproto_timeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
ggproto_timeline
}
\description{
ggproto function for geom_timeline
}
\seealso{
\code{\link{geom_timeline}}
}
\keyword{internal}
