% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\docType{data}
\name{ggproto_timeline_label}
\alias{ggproto_timeline_label}
\title{ggproto function for geom_timeline_label}
\format{
An object of class \code{ggproto_timeline_label} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
ggproto_timeline_label
}
\description{
ggproto function for geom_timeline_label
}
\seealso{
\code{\link{geom_timeline_label}}
}
\keyword{internal}
