% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain and clean data.R
\name{load_data}
\alias{load_data}
\title{Module 1: Obtain and Clean Data}
\usage{
load_data(path = file.path("data_raw", "signif.txt"))
}
\arguments{
\item{path}{A character string file path to the downloaded source.}
}
\value{
Returns a dataframe result after calling readr's read_delim
}
\description{
Take raw data frame and return a clean data frame
}
\examples{
\dontrun{
  df <- load_data()
  print(class(df))
  head(df)
}

}
