% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{make_label}
\alias{make_label}
\title{helper function for eq_create_label}
\usage{
make_label(x, label_name)
}
\description{
helper function for eq_create_label
}
\seealso{
\code{\link{eq_create_label}}
}
