% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_eq_timeline.R
\name{plot_eq_timeline}
\alias{plot_eq_timeline}
\title{Plot noaa timeline}
\usage{
plot_eq_timeline(data, label = LOCATION_NAME)
}
\arguments{
\item{data}{noaa dataset prepared by \code{\link{eq_clean_data}}.}

\item{label}{(\emph{tidy eval syntax}) Column used for labeling.}
}
\description{
Wrapper for \code{\link{geom_timeline}} and  \code{link{geom_timeline_label}}.
Data is grouped by COUNTRY.
}
\examples{
\dontrun{
noaa_data <- system.file("extdata", "noaa_earthquakes.tsv", package = "noaa") \%>\%
  eq_read_data \%>\%
  eq_clean_data \%>\%
  filter(COUNTRY \%in\% c("USA", "CHINA") & YEAR > 2000)

timeline <- noaa_data \%>\%
  plot_eq_timeline(label = DATE)
}

}
\seealso{
\code{\link{geom_timeline}}, \code{\link{geom_timeline_label}}
}
