% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomlabel.R
\name{stat_timelabel}
\alias{stat_timelabel}
\title{Corresponding function to prepare layer object for StatTimelabel annotation.}
\usage{
stat_timelabel(mapping = NULL, data = NULL, geom = "timelabel",
  position = "identity", na.rm = FALSE, n_max = NA, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\value{
layer object for plotting purpose
}
\description{
This is not necessary but prepared for complete functionality. It refers to
geom_timelabel() and StatTimelabel class for plotting purpose. As it refers
geom_timelable(), it means the function can be used 'independent' of
geom_timeline().
}
\examples{
\dontrun{
ggplot(data = selected) +
  aes(x=DATE, y=COUNTRY,
      colour = TOTAL_DEATHS,
      size = EQ_PRIMARY,
      label=LOCATION_NAME, byCol=EQ_PRIMARY,
      xmin = xmin1, xmax = xmax1) +
      geom_timeline() +
  stat_timelabel(n_max=5)+
  theme_timeline() +
  scale_size_continuous(name = "Richter scale value") + #change legend name
  scale_colour_continuous(name = "# Deaths")}
}
