% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomlabel.R
\docType{data}
\name{StatTimelabel}
\alias{StatTimelabel}
\title{Stat class for the annotation of earthquake timeline}
\format{An object of class \code{StatTimelabel} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimelabel
}
\description{
The core work is done in the \code{\link{compute_group}} function, which is
to select data rows bounded by specified dates, and filter data according to
n_max and byCol. REfer \code{\link{compute_timeline_label}} for further
details.
}
\keyword{datasets}
