% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtimeline.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{Stat class for the earthquake data}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 3.}
\usage{
StatTimeline
}
\description{
The core work is done in the \code{\link{compute_group}} function, which is
to select data rows bounded by specified dates.
}
\keyword{datasets}
