% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_geom_functions_week2.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Creates a theme for timeline visualization}
\usage{
theme_timeline()
}
\description{
A simple theme that makes \code{\link{geom_timeline}}
visualize better.
}
\examples{
\dontrun{
data \%>\% eq_clean_data() \%>\%
filter(COUNTRY \%in\% c("GREECE", "ITALY"), YEAR > 2000) \%>\%
   ggplot(aes(x = DATE,
              y = COUNTRY,
              color = as.numeric(TOTAL_DEATHS),
              size = as.numeric(EQ_PRIMARY)
   )) +
   geom_timeline() +
   theme_timeline()
}

}
