% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{draw_timeline}
\alias{draw_timeline}
\title{A helper function for \code{draw_group} geomTimeline}
\usage{
draw_timeline(data, panel_scales, coord)
}
\arguments{
\item{data}{data}

\item{panel_scales}{panel_scales}

\item{coord}{coord}
}
\description{
A helper function for \code{draw_group} geomTimeline
}
