% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_week1.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Cleans raw data from NOAA}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{A data frame with raw data obtained from NOAA website (see below)}
}
\value{
A data frame with cleaned date, latitude, longitude and location
columns
}
\description{
\code{eq_clean_data} cleans date, latitude and longitude, and location name
from NOAA raw data
}
\details{
The function takes raw data from NOAA website as an input
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}. It adds
a column DATE with cleaned date (Date format), transforms LATITUDE and
LONGITUDE columns as numeric objects and transforms LOCATION_NAME by removing
the country and transforming to title case.
}
\examples{
\dontrun{
data <- readr::read_delim("earthquakes.txt", delim = "\t")
clean_data <- eq_clean_data(data)
}

}
