% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtaclean.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean earthquake data}
\usage{
eq_clean_data(frame)
}
\arguments{
\item{frame}{The earthquake data frame. See the documentaton in the data part.}
}
\value{
Data frame like an earthquake data, not incuding
'YEAR','MONTH','DAY','HOUR','MINUTE','SECOND', and added 'data' field.
}
\description{
Creates a single 'date' datetime object out of the fields
'YEAR','MONTH','DAY','HOUR','MINUTE','SECOND'. The 'date' is constructed filling the NA fields
as '1' for month and day, and '0' for the time fields.
}
\examples{
\dontrun{
  earthquake_cleaned <- eq_clean_data(earthquakes)
}
}
