% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean and Format NOAA Data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{The raw dataframe with at least Year, Month, Day, Country, Location_Name, Latitude, Longitiude, EQ_Primary, and TOtal Death fields}
}
\value{
Returns a tbl_df date.frame
}
\description{
This function takes a dataframe and drops date related NAs, converts data types to appropriate type,
removes pre-BC dates due to low information, fixes date parts that do not fit conventionall format,
and returns a cleaned dataframe.
}
\details{
Takes a dataframe of earthquake information and drops date realted NAs, changes the date parts to integers, filters for post-BC dates, fixes dates components to YYYY-MM-DD format, combines date parts into a DATE field, and changes remaing fields to proper type.
}
\examples{
\dontrun{eq_clean_data(df= data)}

}
