% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain and clean data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean the raw NOAA earthquake data}
\usage{
eq_clean_data(df_raw)
}
\arguments{
\item{df_raw}{}
}
\value{
Returns a cleaned dataframe
}
\description{
This function takes the raw NOAA data and cleans it
}
\examples{
\dontrun{
  df <- load_data() \%>\% dplyr::filter(YEAR >= 0) \%>\%
       dplyr::mutate(date = get_date(DAY, MONTH, YEAR))
  print(class(df))
  head(df)
}

}
