% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{a data.frame containing NOAA earthquake data.}
}
\value{
A data.frame with DATE column in POSIXct format and LATITUDE and LONGITUDE column in numeric class.
}
\description{
Adds a DATE column and changes the type of LATITUDE and LONGITUDE column to \code{numeric}.
Also pads zeros to make YEAR columns four digits.
}
\examples{
\dontrun{
noaa.df <- eq_clean_data(readr::read_tsv("data-raw/signif.txt"))
noaa.df <- eq_clean_data(raw.noaa.df)
}

}
