% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Data}
\usage{
eq_clean_data(data)
}
\arguments{
\item{data}{A raw NOAA data frame.}
}
\value{
A clean NOAA data frame.
}
\description{
The function \code{eq_clean_data} cleans a raw NOAA data frame. In more
detail, it adds a date column, converts latitude and longitude columns to
numeric, and cleans the location name column.
}
\examples{
# import libraries
library(dplyr)
library(readr)
# set working directory
setwd(system.file("extdata", package = "noaa"))
# read and clean NOAA data
data <- read_delim("signif.txt", "\\t") \%>\%
   eq_clean_data()

}
