% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA earthquake data}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{A data frame containing raw NOAA earthquake data}
}
\value{
A clean data frame containing date, latitude, longitude, depth, magnitude, and location columns
}
\description{
This function takes a raw NOAA earthquake data frame and returns a clean data frame
with a date column created by uniting the year, month, day, and time columns, and
converting it to the Date class; the LATITUDE and LONGITUDE columns converted to numeric class;
and the LOCATION_NAME column cleaned by stripping out the country name (including the colon)
and converting names to title case (as opposed to all caps).
}
\examples{
\dontrun{
path <- "path/to/your/data.txt"
earthquakes <- read_delim(path, delim = "\t")
earthquakes_clean <- eq_clean_data(earthquakes)
}

}
