% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthquake.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Function to Create Date of Occurrence of Earthquake.}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{Data file on earthquakes from U.S. National Oceanographic and Atmospheric Administration (NOAA).}
}
\value{
An object of class data.frame or tbl_df or tibble.
}
\description{
The goal of the function is to The clean data frame so that it has the following:
(i) A date column created by uniting the year, month, day and converting it to the Date class.
(ii) LATITUDE and LONGITUDE columns converted to numeric class.
}
\examples{
eq_clean_data(eaq_data)

}
