% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA data}
\usage{
eq_clean_data(dataframe)
}
\arguments{
\item{dataframe}{A data.frame including raw NOAA earthquake data, as
read from the original .txt file.}
}
\value{
A tibble with the new column and modified class.
}
\description{
This function takes in raw NOAA earthquake data as a data.frame
(using the original names) and converts the \code{LONGITUDE} and
\code{LATITUDE} columns to
numeric class. It also creates a DATE variable that combines the
\code{YEAR}, \code{MONTH} and \code{DAY} variables to a a Date class object.
}
\examples{
eq_clean_data(NOAA_data)
}
