% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean Earthquake Dataframe}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{raw data frame of earthquake data}
}
\value{
df cleansed version of the earthquake data
}
\description{
The function converts Lat and Lon from character to numeric. For dates, it
pastes the year, month and day as  character and convert it to date. The date
conversion works fine for A.D dates, but for B.C. dates, it needs to be
changed manually by forcing the year to negative. It also converts
TOTAL_DEATHS and EQ_PRIMARY columns to numeric, while changing COUNTRY column
to factor. In addition, it calls \code{\link{eq_location_clean}} to format
LOCATION_NAME column.
}
\examples{
\dontrun{df <- eq_clean_data(df)}
}
