% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_tools.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Given a data frame containing NOAA earthquake data, this function returns a popup-text for each
observation in the data frame. This popup-text consists of location name, magnitude and
total number of deaths. If any of this information is not available, it will not show up in the
popup-text.}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{The data frame from which we will generate the popup-texts.}
}
\value{
A list of popup-texts for the earthquake observations given as input.
}
\description{
Given a data frame containing NOAA earthquake data, this function returns a popup-text for each
observation in the data frame. This popup-text consists of location name, magnitude and
total number of deaths. If any of this information is not available, it will not show up in the
popup-text.
}
\examples{
readr::read_delim(file = system.file("extdata", "signif.txt", package="noaa"), delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}
