% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Add Detailed Annotations to Leaflet Map}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{data to plot}
}
\value{
Returns an interactive leaflet map with detailed annotations when present
}
\description{
This function add date, location, magnitude, and total deaths, in an annotation on an interactive map.
}
\details{
Adds annoations to an interactive map created with eq_map.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
    eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
   dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
   eq_map(annot_col = "popup_text")
}

}
