% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingTools.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Eq Create Label}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data.frame containing NOAA earthquake data.}
}
\value{
A string with LOCATION_NAME, EQ_PRIMARY and TOTAL_DEATHS.
}
\description{
Creates a label with column LOCATION_NAME, EQ_PRIMARY and TOTAL_DEATHS.
}
\examples{
\dontrun{
mexico.df<- mexico.df \%>\% dplyr::mutate(popup_text = eq_create_label(.))
eq_map(mexico.df, "popup_text")
}

}
