% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML Label for Earthquakes}
\usage{
eq_create_label(data, location = "LOCATION_NAME", magnitude = "EQ_PRIMARY",
  total_deaths = "TOTAL_DEATHS")
}
\arguments{
\item{data}{A NOAA data frame.}

\item{location}{The column of the data frame depicting the location of
earthquakes.}

\item{magnitude}{The column of the data frame depicting the magnitude of
earthquakes.}

\item{total_deaths}{The column of the data frame depicitn the number of
total deaths of earthquakes.}
}
\value{
A HTML label containing location, magnitude, and total deaths
information of earthquakes.
}
\description{
The function \code{eq_create_label} creates HTML labels for earthquakes
that can be used as annotation text in function \code{eq_map}.
}
\examples{
# import libraries
library(dplyr)
library(readr)
# set working directory
setwd(system.file("extdata", package = "noaa"))
# read and clean NOAA data
data <- readr::read_delim("signif.txt", "\\t") \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "MEXICO", YEAR > 2000)
# create HTML labels
data <- data \%>\%
   dplyr::mutate(html_label = eq_create_label(.))
# map and annotate NOAA data
eq_map(data, annot_col = "html_label")

}
