% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create popup labels for earthquakes}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{A data frame of earthquake data}
}
\value{
A character vector of popup labels
}
\description{
This function takes in earthquake data and creates popup labels for the map
visualization.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.txt", delim = "\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(country == "MEXICO" & lubridate::year(date) >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
