% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflt_map.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Function to Create Labels for Annotation.}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{An object of class data.frame or tibble or tbl_df that contains data on earthquakes.}
}
\value{
An object of class tibble that comprises the labels.
}
\description{
A function to create labels that comprises total deaths from a earthquake, its location and
magnitude. These labels appear on the pop ups that appear on the map of earthquakes created by eq_map function.
}
\examples{
\dontrun{
library(magrittr)
data(eaq_data)
eaq_data\%>\%
  eq_clean_data()\%>\%
  dplyr::filter(`Country` == "MEXICO" & lubridate::year(Date) >= 2000) \%>\%
  dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
  eq_map(annot_col = "popup_text")
}
}
