% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create labels for an earthquake location map}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{An NOAA earthquake dataframe, with information on locations, magnitude
and death toll of earthquakes.}
}
\value{
An HTML-like column vector with information for the map popups.
}
\description{
Create labels for an earthquake location map
}
\examples{
library(dplyr)
eq_data \%>\% eq_clean_data() \%>\% eq_location_clean() \%>\%
   dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
   mutate(popup_text = eq_create_label(.)) \%>\% eq_map(annot_col = popup_text)

}
