% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivemap.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create HTML formatted pop up text for the earthquake interactive map}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{data frame of the selected earthquake incidents. the data frame has
to have columns for latitude, longitude, LOCATION_NAME, TOAL_DEATHS AND
EQ_PRIMARY.}
}
\value{
vector of characters for HTML formatted popup messages
}
\description{
The function takes in the earthquake data frame containing columns of of:
\itemise{ \item Location name \item Total deaths \item Magnitude of
earthquake } It formats the information to HTML so that it can be used by
\code{\link{eq_map}}.
}
\examples{
\dontrun{df \%>\% dplyr::mutate(popup_text = eq_create_label(.))}
}
