% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_week1.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Cleans earthquake location data}
\usage{
eq_location_clean(raw_data)
}
\arguments{
\item{raw_data}{A data frame with raw data obtained from NOAA website}
}
\value{
A data frame with cleaned LOCATION_NAME column
}
\description{
Cleans earthquake location data
}
\details{
This function transforms NOAA raw data frame's LOCATION_NAME column by
trimming the country name (if applicable) and converting to title case
}
\note{
The function is not exported
}
\examples{
\dontrun{
raw_data <- readr::read_delim("data.gz", delim = "\t")
clean_data <- eq_location_clean(raw_data)
}

}
