% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean and Format Location Data}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{The erthquake data with raw location names}
}
\value{
Returns a tbl_df data.frame
}
\description{
This funtion cleans the LOCATION_NAME column by stripping out the country name (including the colon) and converts names to title case (as opposed to all caps).
}
\details{
This function takes the raw LOCATION_NAME and extracts everything after the ':', then removes the colon, replaces any LOCATION_NAMEs less than 5 characters with the COUNTRY name, and then converts to title case.
}
\examples{
\dontrun{eq_location_clean(df=data)}

}
