% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtain and clean data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Location_NAME column}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{an object data frame}
}
\value{
Returns a cleaned dataframe as described
}
\description{
eq_clean_data () function also needs functionality of clean LOCATION_NAME that is
devoid of country name
}
\examples{
\dontrun{
  df <- load_data() \%>\% eq_location_clean()
  print(class(df))
  head(df)
}

}
