% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyData.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Location Clean}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{a data.frame containing NOAA earthquake data.}
}
\value{
A data.frame with country name removed from LOCATION_NAME column. And the column is converted into Title case.
}
\description{
The country name from LOCATION_NAME column is stripped out. And the column data is converted into Title case.
}
\note{
This function makes a call to \code{eq_clean_data()} function to make the process simple.
}
\examples{
\dontrun{
noaa.df <- eq_location_clean(readr::read_tsv("data-raw/signif.txt"))
noaa.df <- eq_location_clean(raw.noaa.df)
}

}
