% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean Location Name of NOAA Data}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{A raw NOAA data frame.}
}
\value{
A NOAA data frame with a cleaned location name column.
}
\description{
The function \code{eq_location_clean} cleans the location names of
a raw NOAA data frame by stripping out the country name and converting the
location names to title case.
}
\examples{
# import libraries
library(dplyr)
library(readr)
# set working directory
setwd(system.file("extdata", package = "noaa"))
# read and clean NOAA data
data <- read_delim("signif.txt", "\\t") \%>\%
   eq_location_clean()

}
