% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Clean NOAA earthquake location}
\usage{
eq_location_clean(dataframe)
}
\arguments{
\item{dataframe}{A data.frame including raw NOAA earthquake data, as
read from the original .txt file.}
}
\value{
A tibble with the cleaned \code{LOCATION_NAME} column.
}
\description{
This function takes a raw NOAA earthquake data.frame and cleans the
\code{LOCATION_NAME} column, by removing the country name and converting the
string to title case.
}
\examples{
eq_location_clean(NOAA_data)
}
