% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Take out country from Location Name}
\usage{
eq_location_clean(df)
}
\arguments{
\item{df}{data frame of the earthquake data}
}
\value{
df data frame with cleansed LOCATION_NAME
}
\description{
The function uses COUNTRY column as the reference, and take out the country
name from LOCATION_NAME column. NOte that the coutry name has to be exact
match, and it has to followed by ':'. It then converts the result to title
case and wrap the result to width of 25 (for plotting purpose).
}
\examples{
\dontrun{df <- eq_location_clean(df)}
}
