% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{eq_map}
\alias{eq_map}
\title{Create Interactive Map of Earthquake Data}
\usage{
eq_map(df, annot_col = "DATE")
}
\arguments{
\item{df}{data to plot'}

\item{annot_col}{column to use for annotations, defaults to DATE}
}
\value{
Returns an interactive leaflet map
}
\description{
This function plots earthquakes and avaialble information about said earthquakes on an interactive map that will allow the used to click on an earthquake to see the relevant details.
}
\details{
Creates an interactive map showing the location of the earthquakes and thier date.
}
\examples{
\dontrun{
readr::read_delim("earthquakes.tsv.gz", delim = "\\t") \%>\%
    eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000) \%>\%
   eq_map()
}

}
