% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappingTools.R
\name{eq_map}
\alias{eq_map}
\title{Eq Map}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A data.frame containing NOAA earthquake data.}

\item{annot_col}{A column in \code{data} that has to be annotate.}
}
\value{
A interactive leaflet map with location marked based on the data and popup determined by \code{annot_col}.
}
\description{
A interactive map to show earthquakes on it as cirle points whose radius is determined by \code{EQ_PRIMARY} column.
It should also annotate a data  based on the \code{annot_col} value.
}
\examples{
\dontrun{
 mexico.df <- noaa.df  \%>\% dplyr::filter(COUNTRY == "MEXICO" & lubridate::year(DATE) >= 2000)
 eq_map(mexico.df, "DATE")
}

}
