% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{eq_map}
\alias{eq_map}
\title{Map and Annotate Earthquakes}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A NOAA data frame with the columns LONGITUDE and LATITUDE of
the earthquakes.}

\item{annot_col}{The column of the data frame depicting the annotation of
the earthquakes.}
}
\value{
The function outputs a map with epicenters of earthquakes as well
as their annotations to the current graphics device.
}
\description{
The function \code{eq_map} maps the epicenters of earthquakes and annotates
each point with a pop up window.
}
\examples{
# import libraries
library(dplyr)
library(readr)
# set working directory
setwd(system.file("extdata", package = "noaa"))
# read and clean NOAA data
data <- readr::read_delim("signif.txt", "\\t") \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY == "MEXICO", YEAR > 2000)
# map and annotate NOAA data
eq_map(data, annot_col = "DATE")

}
