% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone_functions.R
\name{eq_map}
\alias{eq_map}
\title{Creates a leaflet map with circle markers representing earthquake locations and their associated annotation}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{A data frame containing earthquake data that has been cleaned using the \code{eq_clean_data} function.}

\item{annot_col}{A string representing the name of the column in the \code{data} data frame that should be used for annotation in the popup text.}
}
\value{
A leaflet map with circle markers representing earthquake locations and their associated annotation.
}
\description{
This function takes earthquake data and an annotation column and creates a leaflet map with circle markers at each earthquake location. The annotation column is used to display additional information about each earthquake in the popup text when the user hovers over a marker.
}
\examples{
\dontrun{
data <- read_delim("earthquakes.txt", delim = "\t") \%>\%
        eq_clean_data() \%>\%
        filter(country == "MEXICO" & year(date) >= 2000)
eq_map(data, annot_col = "date")
}
}
