% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactivemap.R
\name{eq_map}
\alias{eq_map}
\title{Create an interactive map zoomed to the data with simple pop-up}
\usage{
eq_map(df, annot_col)
}
\arguments{
\item{df}{data frame of the selected earthquake incidents. the data frame has
to have latitude, longitude and a column for annotation purpose.}

\item{annot_col}{character specifying the column where the annotation text is
displayed.}
}
\description{
The function creates an interactive map to display the location of the
selected earthquakes. When clicked, it displays the date of the earthquake.
}
\examples{
\dontrun{df \%>\% eq_map(df, "DATE")}
}
