% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtaclean.R
\name{eq_read_data}
\alias{eq_read_data}
\title{Read earthquake data}
\usage{
eq_read_data(filename)
}
\arguments{
\item{filename}{The file to process}
}
\value{
data frame, containing the earthquake data
}
\description{
Read a tab separated text file containing the NOAA dataset and returns a data frame.
}
\examples{
eq <- eq_read_data(filename="data-raw/signif.txt.bz2" )

}
