% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_read_data.R
\name{eq_read_data}
\alias{eq_read_data}
\title{<title>}
\usage{
eq_read_data(path)
}
\arguments{
\item{<x>}{(\emph{character}) <description>.}
}
\value{
FIBS data is returned as tibble (see \code{\link[dplyr]{tbl_df}})
}
\description{
<description> \cr
\code{<function name>}.
}
\details{
<details> \code{\link{<other function name>}}
}
\examples{
\dontrun{
...
}

}
\references{
US National Highway Traffic Safety Administration \cr
(\href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars})
}
\seealso{
\code{\link{<other function name>}}
}
