% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\arguments{
\item{...}{other arguments passed on to [layer()]. These are
often aesthetics, used to set an aesthetic to a fixed value, like
`color = "red"` or `size = 3`. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If `FALSE`, the default, missing values are removed with
a warning. If `TRUE`, missing values are silently removed.}
}
\description{
The timeline geom is used to create discrete point plots over a timeline.
Time line of events ranging from xmin to xmax dates with a point for each event.
Optional aesthetics include color, size, and alpha (for transparency).
The xaesthetic is a date and an optional y aesthetic is a factor indicating some
stratification in which case multiple time lines will be plotted.
}
\section{Aesthetics}{

geom_timeline understand the following aesthetics (required aesthetics is 'x'):
\describe{
  \item{x}{datetime of the event}
  \item{size}{the size of the circle drawn at the event}
  \item{colour}{the colour of the circle  drawn at the event}
  \item{y}{free groupiong parameter}
  }
}

\examples{
\dontrun{
  # Use the built-in NOAA dataset filtered for two countries
  eqdta <- eq_location_clean(eq_clean_data(earthquakes)) \%>\%
  filter( date > ymd("20000101") & (COUNTRY=="USA" | COUNTRY=='CHINA' ) ) \%>\%
    group_by(COUNTRY)

  # Draw the geom
  ggplot (data = eqdta) +
    geom_timeline(
      aes(
        x = date,
        y = COUNTRY,
        size = FOCAL_DEPTH,
        colour = EQ_PRIMARY
      )
    )
}

}
\author{
József Varga
}
