% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create Timeline Geom for Earthquake Data}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{defaults to NULL}

\item{data}{defaults to NULL, unless data passed in}

\item{stat}{defautls to 'identity'}

\item{position}{defautls to 'identity'}

\item{na.rm}{defaults to FALSE, do not remove NAs}

\item{show.legend}{defaults to NA}

\item{inherit.aes}{defaults to TRUE}

\item{...}{takes additional arguments}
}
\value{
Returns a ggplot object
}
\description{
This function uses ggplot to visualize a timeline of earthquakes.
}
\details{
Create a new geom called geom_timeline that is used to plot a timline of earthquakes
}
\examples{
\dontrun{data \%>\%
ggplot() +
geom_timeline(aes(x = DATE,
y = COUNTRY,
fill = TOTAL_DEATHS,
size = EQ_PRIMARY))
}

}
