% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization tools.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Module 2: Visualization tools: geom_timeline()}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Building geom_timeline()
a geom for ggplot2 for plotting a timeline of earthquakes (xmin to xmax of dates) for each earthquake
points (in the example) represent earthquake events, point size indicating
earthquake magnitude and colour represent number of deaths.
xaes = the date (required), yaes = country (optional)
}
\examples{
\dontrun{
 df <- df \%>\% filter(COUNTRY \%in\% c("CHINA", "USA"), YEAR > 2000)
 ggplot(df, aes(x = date, y = COUNTRY,
                color = as.numeric(TOTAL_DEATHS),
                size = as.numeric(EQ_PRIMARY),
                label = CLEAN_LOCATION_NAME)) +
   geom_timeline() +
   labs(size = "Richter scale value", color = "# deaths") +
   ggplot2::theme(panel.background = ggplot2::element_blank(),
         legend.position = "bottom",
         axis.title.y = ggplot2::element_blank()) +
   ggplot2::xlab("DATE")
}

}
