% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Geom timeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\value{
ggplot2 object to visualize the timeline and magnitude of earthquakes from NOAA data.
}
\description{
A custom geom to plot the timeline of earthquake with magnitude and number of death.
}
