% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Plot Timeline Geom}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{Boolean indicating wheather to remove NA values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{Additional optional parameters.}
}
\value{
The function \code{geom_timeline} plots a timeline geom.
}
\description{
The function \code{geom_timeline} plots a timeline geom showing one or
multiple timelines of earthquakes.
}
\section{Aesthetics}{

\code{x}: The column of the given data frame depicting the x values of the
timeline geom, usually date. \cr
\code{y}: Optional. The column of the given data frame depicting the y
values of the timeline geom, usually location. \cr
\code{size}: Optional. The column of the given data frame depicting the size
of each point in the timeline geom. \cr
\code{colour}: Optional.  The column of the given data frame depicting the
color of each point in the timeline geom. \cr
\code{fill}: Optional. A color string depicting the default fill of each
point in the timeline geom. \cr
\code{shape}: Optional. The column of the given data frame depicting the
shape of each point in the timeline geom. \cr
\code{alpha}: Optional. The column of the given data frame depicting the
alpha value of each point in the timeline geom. \cr
\code{stroke}: Optional. The stroke of the image. \cr
}

\examples{
# import libraries
library(dplyr)
library(readr)
# set working directory
setwd(system.file("extdata", package = "noaa"))
# read and clean NOAA data
data <- read_delim("signif.txt", "\\t") \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000)
# plot NOAA data
ggplot2::ggplot(data, ggplot2::aes(DATE,
                                   colour = as.numeric(TOTAL_DEATHS),
                                   y = COUNTRY,
                                   size = as.numeric(EQ_PRIMARY))) +
   geom_timeline() +
   ggplot2::guides(size = ggplot2::guide_legend(title = "Richter Scale")) +
   ggplot2::scale_colour_continuous(name = "Number of Deaths") +
   ggplot2::theme_classic()

}
