% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone_functions.R
\name{geom_timeline}
\alias{geom_timeline}
\title{GeomTimeline for ggplot2}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping.}

\item{data}{The dataset to be used for the plot.}

\item{stat}{The statistical transformation to use on the data for this layer.}

\item{position}{Position adjustment to use for this layer.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.}

\item{...}{Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like color = "red" or size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
A ggplot layer containing a time line of earthquakes.
}
\description{
GeomTimeline is a custom geom for ggplot2 that allows you to plot a timeline of data points, where the x-axis represents time and the y-axis is fixed at y=0. Optional aesthetics include size, color, alpha, shape, stroke, and fill. The default values for aesthetics are size=1, color='grey50', alpha=0.5, shape=19, stroke=0.5, and fill=NA.
}
\examples{
\dontrun{
ggplot(data = earthquakes) +
geom_timeline(mapping = aes(x = date, size = depth, color = magnitude)) +
scale_size_continuous(trans = 'reverse')
}

}
