% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtimeline.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Corresponding function to GeomTimeline}
\usage{
geom_timeline(mapping = NULL, data = NULL, stat = "timeline",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\value{
layer object for plotting purpose
}
\description{
The function is used with ggplot2 to produce plot for the earthquake data.
For input arguments, please refer to \code{geom_point} for further details.
}
\examples{
\dontrun{ggplot(data = selected) +
   aes(x=DATE, colour = TOTAL_DEATHS,
   size = EQ_PRIMARY) +
   geom_timeline(xmin=xmin, xmax=xmax) }
}
