% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_label.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Create Timeline Lablels}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{defaults to NULL}

\item{data}{defaults to NULL, unless data passed in}

\item{stat}{defautls to 'identity'}

\item{position}{defautls to 'identity'}

\item{na.rm}{defaults to FALSE, do not remove NAs}

\item{show.legend}{defaults to NA}

\item{inherit.aes}{defaults to TRUE}

\item{...}{takes additional arguments}
}
\value{
Returns a ggplot object
}
\description{
This function creates a new geom that  can build on geom_timeline and add labels to the top n biggest earthqakes by magnitude
}
\details{
Create a new geom called geom_timeline_label that is used to plot relevant top_n labels on geom_timeline
}
\examples{
\dontrun{data \%>\%
    ggplot() +
    geom_timeline(aes(
        x = DATE,
        y = COUNTRY,
        fill = TOTAL_DEATHS,
        size = EQ_PRIMARY)) +
    geom_timeline_label(aes(
        x = DATE,
        y = COUNTRY,
        magnitude = EQ_PRIMARY,
        label = LOCATION_NAME
        ,n_max=3 #number of lables to show))
}

}
