% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization tools.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Visualization tool: geom_timeline_label}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n_max = 5,
  ...
)
}
\arguments{
\item{n_max}{An integer (default = 5) max # of earthquakes}
}
\description{
Scripting geom_timeline_label function
geom function aims to take nmax value of earthquakes as vertical line label
}
\examples{
\dontrun{
ggplot(df, aes(x = Date, y = COUNTRY,
               color = as.numeric(TOTAL_DEATHS),
               size = as.numeric(EQ_PRIMARY),
               label = CLEAN_LOCATION_NAME)) +
  geom_timeline() +
  labs(size = "Richter scale value", color = "# deaths") +
  ggplot2::theme(panel.background = ggplot2::element_blank(),
        legend.position = "bottom",
        axis.title.y = ggplot2::element_blank()) + ggplot2::xlab("DATE") +
  geom_timeline_label(data=df)
}
}
