% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelineLabel.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Geom Timeline Label}
\usage{
geom_timeline_label(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{na.rm}{na.rm}

\item{show.legend}{show.legend}

\item{inherit.aes}{inherit.aes}

\item{...}{...}
}
\value{
a ggplot2 object which can label earthquakes created by \code{geom_timeline}.
}
\description{
A geom to add label to the earthquakes.
It should have an option \code{n_max} to select the n maximum earthquakes ordered by magnitude.
}
