% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone_functions.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{GeomTimeLineLabel geom for label to earthquake data plot}
\format{
A ggplot2::ggproto object
}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  na.rm = TRUE,
  show.legend = NA,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{A set of aesthetic mappings created by aes() or aes_(). Only x and tags are required.}

\item{data}{A data.frame or tibble containing the variables to be plotted.}

\item{na.rm}{A logical value. Should missing values be removed from the plot?}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped to variables with guide boxes.}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{The position adjustment to use for overlapping points or bars, as a string.}

\item{inherit.aes}{A logical value. Should the default aesthetics be used for this layer? If FALSE, overrides are used instead.}

\item{...}{Additional arguments passed to layer().}
}
\value{
A ggplot2 plot with a timeline label.
}
\description{
This geom adds a vertical line to each data point with a text annotation attached to each line.
There is an option to subset to n_max number of earthquakes, where the n_max largest (by magnitude) earthquakes are selected.
Aesthetics are x, which is the date of the earthquake, and tags which is the column name from which annotations will be obtained.
}
\examples{
\dontrun{
library(ggplot2)

# Create earthquake data
data <- quakes[order(quakes$mag, decreasing = TRUE), ]
}
}
