% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_geom.R
\name{geom_timeline_label}
\alias{geom_timeline_label}
\title{Timeline of Earthquakes (annoted).}
\usage{
geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \href{https://www.rdocumentation.org/packages/ggplot2/versions/3.3.3/topics/aes}{aes()} or \href{https://www.rdocumentation.org/packages/ggplot2/versions/3.3.3/topics/aes_}{aes_()}.
If specified and `inherit.aes = TRUE`` (the default), it is combined with the default mapping at the top
level of the plot. You must supply mapping if there is no plot mapping}

\item{data}{The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to \href{https://www.rdocumentation.org/packages/ggplot2/versions/3.3.3/topics/ggplot}{ggplot()}.

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a
data frame. See \href{https://www.rdocumentation.org/packages/precrec/versions/0.12.5/topics/fortify}{fortify()} for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame,
and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).}

\item{stat}{The statistical transformation to use on the data for this layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of a call to a position adjustment function.}

\item{na.rm}{If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends? NA, the default, includes if
any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to
finely select the aesthetics to display.}

\item{inherit.aes}{If FALSE, overrides the default aesthetics, rather than combining with them.
This is most useful for helper functions that define both data and aesthetics and shouldn't inherit
behaviour from the default plot specification, e.g. \href{https://www.rdocumentation.org/packages/ggplot2/versions/0.9.0/topics/borders}{borders()}.}

\item{...}{Other arguments passed on to \href{https://ggplot2.tidyverse.org/reference/layer.html}{layer()}. These are often aesthetics, used to set an aesthetic to a
fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.}
}
\value{
A plot of the timeline of earthquakes with annotations.
}
\description{
A geom for adding annotations to the earthquake data. This geom adds a vertical line to each
data point with a text annotation (e.g. the location of the earthquake) attached to each line.
There should be an option to subset to n_max number of earthquakes, where we take the n_max largest
(by magnitude) earthquakes. Aesthetics are x, which is the date of the earthquake and label which takes
the column name from which annotations will be obtained.
}
\examples{
\dontrun{
library(magrittr)
data(eaq_data)
eaq_data\%>\%
  eq_clean_data()\%>\%
  dplyr::filter(`Country` \%in\% c("MEXICO") & !(is.na(Mag))) \%>\%
  ggplot(aes(x = Year, y = Country, size = Mag, colour = `Total Deaths`)) +
   geom_timeline_label(alpha = 0.3, label = `Location Name`)
}
}
