% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tools.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Title}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\arguments{
\item{coord}{}
}
\description{
Title
}
\keyword{datasets}
