% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_tools.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Function used internally to construct grid objects for the layer constructed by "geom_timeline".}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\value{
A grid object list ready to be added to a ggplot layer.
}
\description{
Function used internally to construct grid objects for the layer constructed by "geom_timeline".
}
\keyword{datasets}
