% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{geomTimeline}
\alias{geomTimeline}
\title{geomTimeline}
\format{An object of class \code{geomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomTimeline
}
\description{
We are using ggproto to create a new class.
}
\keyword{datasets}
