% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Construct new Class for Timeline Geom}
\format{An object of class \code{GeomTimeline} (inherits from \code{Geom}, \code{ggproto}) of length 5.}
\usage{
GeomTimeline
}
\value{
A new ggproto object for a timeline geom.
}
\description{
\code{GeomTimeline} constructs a new class for a timeline geom plotting one
or multiple timelines of earthquakes.
}
\section{Aesthetics}{

\code{x}: The column of the given data frame depicting the x values of the
timeline geom, usually date. \cr
\code{y}: Optional. The column of the given data frame depicting the y
values of the timeline geom, usually location. \cr
\code{size}: Optional. The column of the given data frame depicting the size
of each point in the timeline geom. \cr
\code{colour}: Optional.  The column of the given data frame depicting the
color of each point in the timeline geom. \cr
\code{fill}: Optional. A color string depicting the default fill of each
point in the timeline geom. \cr
\code{shape}: Optional. The column of the given data frame depicting the
shape of each point in the timeline geom. \cr
\code{alpha}: Optional. The column of the given data frame depicting the
alpha value of each point in the timeline geom. \cr
\code{stroke}: Optional. The stroke of the timeline geom. \cr
}

\examples{
# import libraries
library(dplyr)
library(readr)
# set working directory
setwd(system.file("extdata", package = "noaa"))
# read and clean NOAA data
data <- read_delim("signif.txt", "\\t") \%>\%
   eq_clean_data() \%>\%
   dplyr::filter(COUNTRY \%in\% c("USA", "CHINA"), YEAR > 2000)
# plot NOAA data
ggplot2::ggplot(data, ggplot2::aes(DATE,
                                   colour = as.numeric(TOTAL_DEATHS),
                                   y = COUNTRY,
                                   size = as.numeric(EQ_PRIMARY))) +
   geom_timeline() +
   ggplot2::guides(size = ggplot2::guide_legend(title = "Richter Scale")) +
   ggplot2::scale_colour_continuous(name = "Number of Deaths") +
   ggplot2::theme_classic()

}
\keyword{datasets}
