% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone_functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline for plotting earthquake data.}
\format{
ggplot2::ggproto
}
\usage{
GeomTimeline
}
\description{
GeomTimeline is a custom geom for ggplot2 that allows you to plot a timeline of data points,
where the x-axis represents time and the y-axis is fixed at y=0. Optional aesthetics include size,
color, alpha, shape, stroke, and fill. The default values for aesthetics are size=1,
color='grey50', alpha=0.5, shape=19, stroke=0.5, and fill=NA.
}
\keyword{datasets}
