% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelineLabel.R
\docType{data}
\name{geomTimelineLabel}
\alias{geomTimelineLabel}
\title{geomTimelineLabel}
\format{An object of class \code{geomTimelineLable} (inherits from \code{Geom}, \code{ggproto}) of length 6.}
\usage{
geomTimelineLabel
}
\description{
geomTimelineLabel
}
\examples{
\dontrun{
get_timeline(noaa.df, c("CHINA", "INDIA"),"2000-01-01","2010-01-01") +
geom_timeline_label(ggplot2::aes(x=DATE, location=LOCATION_NAME,xmin=xmin,xmax=xmax,size=EQ_PRIMARY,y=COUNTRY), n_max = 5)
}

}
\keyword{datasets}
