% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstone_functions.R
\docType{data}
\name{GeomTimeLineLabel}
\alias{GeomTimeLineLabel}
\title{GeomTimeLineLabel geom for adding annotations to earthquake data}
\format{
ggplot2::ggproto
}
\usage{
GeomTimeLineLabel
}
\description{
This geom adds a vertical line to each data point with a text annotation attached to each line.
There is an option to subset to n_max number of earthquakes, where the n_max largest (by magnitude) earthquakes are selected.
Aesthetics are x, which is the date of the earthquake, and tags which is the column name from which annotations will be obtained.
}
\keyword{datasets}
