% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{StatTimeline}
\format{An object of class \code{StatTimeline} (inherits from \code{Stat}, \code{ggproto}) of length 4.}
\usage{
StatTimeline
}
\description{
StatTimeline
}
\keyword{datasets}
