% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomtimeline.R
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for the earthquake data}
\usage{
theme_timeline()
}
\value{
a theme object to be attached to ggplot command
}
\description{
A few tweaks to turn off the background and grids.
}
\examples{
\dontrun{ggplot(data = selected) +
aes(x=DATE, y=COUNTRY,
   colour = TOTAL_DEATHS,
   size = EQ_PRIMARY) +
   geom_timeline(xmin=xmin, xmax=xmax) +
   theme_timeline() +
   scale_size_continuous(name = "Richter scale value") + #change legend name
   scale_colour_continuous(name = "# Deaths") +
   scale_x_date(limits = c(xmin-months(6), xmax+months(6))) #extend the axis}
}
